requirements.txt
langgraph>=0.2.28
langchain>=0.3.7
openai>=1.51.0
python-dotenv>=1.0.1
pydantic>=2.9.2
httpx>=0.27.2
tiktoken>=0.7.0
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
numpy>=1.26.4
pandas>=2.2.2
trafilatura>=1.8.0
beautifulsoup4>=4.12.3
readability-lxml>=0.8.1
duckduckgo-search>=6.2.6
tavily-python>=0.3.3
rapidfuzz>=3.9.7
tenacity>=9.0.0
fastapi>=0.114.2
uvicorn>=0.30.6
gradio>=4.44.0
networkx>=3.3

python-dotenv
openai>=1.40.0
requests
beautifulsoup4
lxml
numpy
gradio>=4.0.0
langgraph>=0.2.0
pytest
